Fix: external storage problem NavRing applications
If the user binds an app which is located on the
external sdcard and is rebooting the button icon
will be left blank because the navring is already
constructed before the sdcard is mounted. To avoid
this we listen to the app available intents.
If it is not available the ring target shows the none icon.
 When available it shows the
correct app icon. So that the user can clearly see
this app is accessable or not. As well when the
user goes into mass storage mode or does
unmount the sdcard or whatever case could
happen that the application is not there.

+ code cleanup who ever did that crap :p

+ to be more consistent to all other features we have an
empty action is now as well shown as an empty action on
the ring

Change-Id: Idce4c28dfc080bd1147fd5b80154d3a50d01141d

Conflicts:
	packages/SystemUI/src/com/android/systemui/SearchPanelView.java